+++ playgrounds/solid/src/routes/login.tsx
@@ -1,11 +1,4 @@
-import {
-  createForm,
-  email,
-  Field,
-  Form,
-  minLength,
-  required,
-} from '@modular-forms/solid';
+import { createForm, email, minLength, required } from '@modular-forms/solid';
 import { FormFooter, FormHeader, TextInput, Title } from '~/components';
 
 type LoginForm = {
@@ -15,7 +8,7 @@ type LoginForm = {
 
 export default function LoginPage() {
   // Create login form
-  const loginForm = createForm<LoginForm>();
+  const [loginForm, { Form, Field }] = createForm<LoginForm>();
 
   return (
     <>
@@ -23,42 +16,42 @@ export default function LoginPage() {
 
       <Form
         class="space-y-12 md:space-y-14 lg:space-y-16"
-        of={loginForm}
         onSubmit={(values) => alert(JSON.stringify(values, null, 4))}
       >
         <FormHeader of={loginForm} heading="Login form" />
         <div class="space-y-8 md:space-y-10 lg:space-y-12">
           <Field
-            of={loginForm}
             name="email"
             validate={[
               required('Please enter your email.'),
               email('The email address is badly formatted.'),
             ]}
           >
-            {(field) => (
-              <TextInput
-                {...field.props}
-                value={field.value}
-                error={field.error}
-                type="email"
-                label="Email"
-                placeholder="example@email.com"
-                required
-              />
-            )}
+            {(field, props) => {
+              console.log('Field children', field.name);
+              return (
+                <TextInput
+                  {...props}
+                  value={field.value}
+                  error={field.error}
+                  type="email"
+                  label="Email"
+                  placeholder="example@email.com"
+                  required
+                />
+              );
+            }}
           </Field>
           <Field
-            of={loginForm}
             name="password"
             validate={[
               required('Please enter your password.'),
               minLength(8, 'You password must have 8 characters or more.'),
             ]}
           >
-            {(field) => (
+            {(field, props) => (
               <TextInput
-                {...field.props}
+                {...props}
                 value={field.value}
                 error={field.error}
                 type="password"