+++ playgrounds/solid/src/routes/special.tsx
@@ -1,4 +1,4 @@
-import { createForm, Field, Form } from '@modular-forms/solid';
+import { createForm } from '@modular-forms/solid';
 import { For } from 'solid-js';
 import {
   FormHeader,
@@ -23,14 +23,14 @@ type SpecialForm = {
     string: string;
   };
   file: {
-    list: FileList;
+    list: File[];
     item: File;
   };
 };
 
 export default function SpecialPage() {
   // Create special form
-  const specialForm = createForm<SpecialForm>();
+  const [specialForm, { Form, Field }] = createForm<SpecialForm>();
 
   return (
     <>
@@ -38,15 +38,14 @@ export default function SpecialPage() {
 
       <Form
         class="space-y-12 md:space-y-14 lg:space-y-16"
-        of={specialForm}
         onSubmit={(values) => alert(JSON.stringify(values, null, 4))}
       >
         <FormHeader of={specialForm} heading="Special form" />
         <div class="space-y-8 md:space-y-10 lg:space-y-12">
-          <Field of={specialForm} name="number">
-            {(field) => (
+          <Field name="number" type="number">
+            {(field, props) => (
               <TextInput
-                {...field.props}
+                {...props}
                 value={field.value}
                 error={field.error}
                 type="number"
@@ -55,10 +54,10 @@ export default function SpecialPage() {
             )}
           </Field>
 
-          <Field of={specialForm} name="range">
-            {(field) => (
+          <Field name="range" type="number">
+            {(field, props) => (
               <Slider
-                {...field.props}
+                {...props}
                 value={field.value}
                 error={field.error}
                 label="Range"
@@ -79,10 +78,10 @@ export default function SpecialPage() {
               ]}
             >
               {({ label, value }) => (
-                <Field of={specialForm} name="checkbox.array">
-                  {(field) => (
+                <Field name="checkbox.array" type="string[]">
+                  {(field, props) => (
                     <Checkbox
-                      {...field.props}
+                      {...props}
                       label={label}
                       value={value}
                       checked={field.value?.includes(value)}
@@ -95,10 +94,10 @@ export default function SpecialPage() {
             </For>
           </div>
 
-          <Field of={specialForm} name="checkbox.boolean">
-            {(field) => (
+          <Field name="checkbox.boolean" type="boolean">
+            {(field, props) => (
               <Checkbox
-                {...field.props}
+                {...props}
                 checked={field.value}
                 error={field.error}
                 label="Checkbox boolean"
@@ -106,10 +105,10 @@ export default function SpecialPage() {
             )}
           </Field>
 
-          <Field of={specialForm} name="select.array">
-            {(field) => (
+          <Field name="select.array" type="string[]">
+            {(field, props) => (
               <Select
-                {...field.props}
+                {...props}
                 value={field.value}
                 options={[
                   { label: 'Option 1', value: 'option_1' },
@@ -123,10 +122,10 @@ export default function SpecialPage() {
             )}
           </Field>
 
-          <Field of={specialForm} name="select.string">
-            {(field) => (
+          <Field name="select.string">
+            {(field, props) => (
               <Select
-                {...field.props}
+                {...props}
                 value={field.value}
                 options={[
                   { label: 'Option 1', value: 'option_1' },
@@ -139,10 +138,10 @@ export default function SpecialPage() {
             )}
           </Field>
 
-          <Field of={specialForm} name="file.list">
-            {(field) => (
+          <Field name="file.list" type="File[]">
+            {(field, props) => (
               <FileInput
-                {...field.props}
+                {...props}
                 value={field.value}
                 error={field.error}
                 label="File list"
@@ -151,10 +150,10 @@ export default function SpecialPage() {
             )}
           </Field>
 
-          <Field of={specialForm} name="file.item">
-            {(field) => (
+          <Field name="file.item" type="File">
+            {(field, props) => (
               <FileInput
-                {...field.props}
+                {...props}
                 value={field.value}
                 error={field.error}
                 label="File item"