+++ website/src/routes/(layout)/playground/login.tsx
@@ -1,11 +1,4 @@
-import {
-  createForm,
-  email,
-  Field,
-  Form,
-  minLength,
-  required,
-} from '@modular-forms/solid';
+import { createForm, email, minLength, required } from '@modular-forms/solid';
 import { onMount } from 'solid-js';
 import { FormFooter, FormHeader, TextInput, Title } from '~/components';
 import { useForm } from '~/contexts';
@@ -17,7 +10,7 @@ type LoginForm = {
 
 export default function LoginPage() {
   // Create login form
-  const loginForm = createForm<LoginForm>();
+  const [loginForm, { Form, Field }] = createForm<LoginForm>();
 
   // Set login form in form context
   onMount(() => useForm().set(loginForm));
@@ -28,22 +21,20 @@ export default function LoginPage() {
 
       <Form
         class="space-y-12 md:space-y-14 lg:space-y-16"
-        of={loginForm}
         onSubmit={(values) => alert(JSON.stringify(values, null, 4))}
       >
         <FormHeader of={loginForm} heading="Login form" />
         <div class="space-y-8 md:space-y-10 lg:space-y-12">
           <Field
-            of={loginForm}
             name="email"
             validate={[
               required('Please enter your email.'),
               email('The email address is badly formatted.'),
             ]}
           >
-            {(field) => (
+            {(field, props) => (
               <TextInput
-                {...field.props}
+                {...props}
                 value={field.value}
                 error={field.error}
                 type="email"
@@ -54,16 +45,15 @@ export default function LoginPage() {
             )}
           </Field>
           <Field
-            of={loginForm}
             name="password"
             validate={[
               required('Please enter your password.'),
               minLength(8, 'You password must have 8 characters or more.'),
             ]}
           >
-            {(field) => (
+            {(field, props) => (
               <TextInput
-                {...field.props}
+                {...props}
                 value={field.value}
                 error={field.error}
                 type="password"