+++ packages/solid/CHANGELOG.md
@@ -4,7 +4,14 @@ All notable changes to the library will be documented in this file.
 
 ## vX.X.X (Month DD, YYYY)
 
-> Note: The package has been revised and refactored and it would be too complicated to mention every change here. Please create an [issue](https://github.com/fabian-hiller/modular-forms/issues/new) if you encounter problems.
+> Note: The package has been revised and refactored and it would be too complicated to mention every change here. Please look at the [release notes](https://github.com/fabian-hiller/modular-forms/releases/tag/vX.X.X-solid) and create an [issue](https://github.com/fabian-hiller/modular-forms/issues/new) if you encounter problems.
+
+- Rename `createForm` primitive to `createFormStore`
+- Add `createForm` primitive with linked components
+- Change `children` property of `Field` component
+- Add `type` property to `Field` component
+- Change behaviour of controlled `<input type="number" />`
+- Make `value` required on `insert` and `replace` method
 
 ## v0.11.0 (March 16, 2023)